+++ modules/Scheduling/AddDrop.php
@@ -1,4 +1,5 @@
 <?php
+require_once 'ProgramFunctions/TipMessage.fnc.php';
 
 DrawHeader( ProgramTitle() );
 
@@ -34,7 +35,49 @@ AND cp.COURSE_ID=c.COURSE_ID
 AND s.STUDENT_ID=se.STUDENT_ID
 AND se.SCHOOL_ID='" . UserSchool() . "'
 AND se.END_DATE BETWEEN '" . $start_date . "' AND '" . $end_date . "'
-ORDER BY DATE DESC", [ 'START_DATE' => 'ProperDate', 'END_DATE' => 'ProperDate' ] );
+ORDER BY DATE DESC", [
+	'FULL_NAME' => '_makeStudentScheduleLink',
+	'START_DATE' => 'ProperDate',
+	'END_DATE' => 'ProperDate'
+] );
 
-$columns = [ 'FULL_NAME' => _( 'Student' ), 'STUDENT_ID' => sprintf( _( '%s ID' ), Config( 'NAME' ) ), 'COURSE_TITLE' => _( 'Course' ), 'TITLE' => _( 'Course Period' ), 'START_DATE' => _( 'Enrolled' ), 'END_DATE' => _( 'Dropped' ) ];
+$columns = [
+	'FULL_NAME' => _( 'Student' ),
+	'STUDENT_ID' => sprintf( _( '%s ID' ), Config( 'NAME' ) ),
+	'COURSE_TITLE' => _( 'Course' ),
+	'TITLE' => _( 'Course Period' ),
+	'START_DATE' => _( 'Enrolled' ),
+	'END_DATE' => _( 'Dropped' ),
+];
+
 ListOutput( $enrollment_RET, $columns, 'Schedule Record', 'Schedule Records' );
+
+/**
+ * Make Student Schedule link
+ *
+ * @since 10.0
+ *
+ * Local function
+ * DBGet() callback
+ *
+ * @param  string $value  Student Full Name.
+ * @param  string $column Column.
+ *
+ * @return string         Link to Student Info program.
+ */
+function _makeStudentScheduleLink( $value, $column = 'FULL_NAME' )
+{
+	global $THIS_RET;
+
+	$modname = 'Scheduling/Schedule.php';
+
+	if ( ! AllowUse( $modname )
+		|| ! $THIS_RET['STUDENT_ID'] )
+	{
+		return $value;
+	}
+
+	$link = 'Modules.php?modname=' . $modname . '&student_id=' . $THIS_RET['STUDENT_ID'];
+
+	return '<a href="' . URLEscape( $link ) . '">' . MakeStudentPhotoTipMessage( $THIS_RET['STUDENT_ID'], $value ) . '</a>';
+}